fix(ci): python-coverage-shards (8) line limit — merge_loop.py and test_pr_monitor_merge_failures.py exceeded 1500 lines
The attention-marker regression commits grew
src/awf/runtime/pr_monitor_runner/merge_loop.py to 1538 lines and
tests/unit/runtime/test_pr_monitor_merge_failures.py to 1673 lines, both
tripping test_first_party_code_files_stay_under_line_limit (MAX=1500).

test split: extract the #661/#663/PRRT_kwDOSJAM6s6La_SZ/LcfXk
attention-marker tests into a new test_pr_monitor_merge_attention.py
module (618 lines). The merge-failure/notification tests stay in the
original file (1102 lines). Behavior unchanged — same fixtures, same
assertions, just relocated; all 5 moved tests and all 18 remaining tests
pass.

merge_loop.py: comment-only condensation of the verbose rationale blocks
added by the recent attention-marker fixes (entry-timestamp reference, the
three non-human-gate clear arms, the atomic marker/attention pairing). No
code or logic changed; all audit-trail references (PRRT IDs / issue
numbers) preserved. File back to 1499 lines.
